% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collin-package.R
\docType{package}
\name{collin-package}
\alias{collin}
\alias{collin-package}
\title{collin: Visualization the Effects of Collinearity in Distributed Lag Models and Other Linear Models}
\description{
Tool to assessing whether the results of a study could be influenced by collinearity. Simulations under a given hypothesized truth regarding effects of an exposure on the outcome are used and the resulting curves of lagged effects are visualized. A user's manual is provided, which includes detailed examples (e.g. a cohort study looking for windows of vulnerability to air pollution, a time series study examining the linear association of air pollution with hospital admissions, and a time series study examining the non-linear association between temperature and mortality). The methods are described in Basagana and Barrera-Gomez (2021) \doi{10.1093/ije/dyab179}.
}
\author{
\strong{Maintainer}: Jose Barrera-Gomez \email{jose.barrera@isglobal.org} (\href{https://orcid.org/0000-0002-2688-6036}{ORCID})

Authors:
\itemize{
  \item Xavier Basagana \email{xavier.basagana@isglobal.org} (\href{https://orcid.org/0000-0002-8457-1489}{ORCID})
}

}
\keyword{internal}
