% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collin-package.R
\docType{package}
\name{collin}
\alias{collin}
\title{collin: Visualizing the effects of collinearity in distributed lag models
  and other linear models}
\description{
Tool to assessing whether the results of a study could be driven
  by collinearity.
}
\section{Procedure}{
 The tool consists of a two-step procedure.

  In the first step, the user provides the fitted model to be analyzed and a
  hypothetical effect pattern of the predictor of interest on the outcome.
  Then, simulations are performed under such a hypothetical effects and the
  given model.

  In the second step, a specific \code{plot} method is used to visualize
  the results of the simulations to assessing whether these results are
  consistent with the results using the real data, so that unexpected results
  could be driven by collinearity.
}

\section{Authors}{

  Jose Barrera-Gomez and Xavier Basagana
  Maintainer: Jose Barrera-Gomez <jose.barrera@isglobal.org>
}

\references{
The methodology used in the package is described in

Basagaña X, Barrera-Gómez J. Visualizing the effects of collinearity in
  distributed lag models. \emph{International Journal of Epidemiology}. 2021;
  51(1):334-344.
}
