% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagpad.R
\name{lagpad}
\alias{lagpad}
\title{Function to get a vector lagged.}
\usage{
lagpad(x, k)
}
\arguments{
\item{x}{numeric vector to be lagged.}

\item{k}{number of lags to be applied.}
}
\value{
A lagged vector.
}
\description{
Function to get a vector lagged.
}
\examples{
 s <- 1:5
 s1 <- lagpad(s, k = 1)
 s2 <- lagpad(s, k = 2)
 data.frame(s, s1, s2)
}
