% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcollindlnm.R
\name{plot.collindlnmlinear}
\alias{plot.collindlnmlinear}
\title{Visualizes effects of collinearity in distributed lag model under an
  hypothetical linear effect pattern.}
\usage{
\method{plot}{collindlnmlinear}(x, lags = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{"collindlnmlinear"}, which is generated by
the function \code{\link{collindlnm}} with \code{shape = "linear"}.}

\item{lags}{a number or a numeric vector indicating at what lags the results
will be visualized. Default (\code{NULL}) shows all lags.}

\item{...}{other parameters to be passed through to plot function.}
}
\value{
A plot showing a comparison between the results under the fitted
  model and the results under the hypothetical true effect.
}
\description{
Visualize the results from a distributed lag model under an hypothetical
  linear effect pattern provided by the user, generated using the function
  \code{\link{collindlnm}}.
}
\examples{
# For detailed examples:
browseVignettes("collin")
}
\seealso{
\code{\link{collindlnm}}, \code{\link{plot.collindlnmnonlinear}}.
}
