% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rbinomod}
\alias{rbinomod}
\title{Internal function to generate random numbers following over-dispersed
  Bernoulli distribution, based in the BetaBinomial distribution. It
  currently does not work and standard Bernoulli distribution is actually
  used instead.}
\usage{
rbinomod(n, prob, d = 1)
}
\arguments{
\item{n}{integer. Number of observations.}

\item{prob}{numeric. Probability of success on each trial.}

\item{d}{numeric. Dispersion parameter.}
}
\value{
a vector with the simulated sample.
}
\description{
Internal function to generate random numbers following over-dispersed
  Bernoulli distribution, based in the BetaBinomial distribution. It
  currently does not work and standard Bernoulli distribution is actually
  used instead.
}
\keyword{internal}
