% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rod}
\alias{rod}
\title{Internal function to generate random numbers following over-dispersed
Bernoulli or Poisson distribution. It wraps \code{rpoisod} and
\code{rbinomod} functions.}
\usage{
rod(n, mu, d, type = c("binomial", "quasibinomial", "poisson", "quasipoisson"))
}
\arguments{
\item{n}{integer. Number of observations.}

\item{mu}{numeric. Probability of success on each trial or mean, for
(quasi)binomial and (quasi)Poisson, respectively.}

\item{d}{numeric. Dispersion parameter.}

\item{type}{character indicating the distribution.}
}
\value{
a vector with the simulated sample.
}
\description{
Internal function to generate random numbers following over-dispersed
Bernoulli or Poisson distribution. It wraps \code{rpoisod} and
\code{rbinomod} functions.
}
\keyword{internal}
