% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{rpoisod}
\alias{rpoisod}
\title{Internal function to generate random numbers following over-dispersed
Poisson distribution based on simple cheat of using a standard negative
binomial, but choosing the scale parameter to give the desired mean vs
variance ratio at the given value of the mean.
Taken from: https://stat.ethz.ch/pipermail/r-help/2002-June/022425.html}
\usage{
rpoisod(n, lambda, d = 1)
}
\arguments{
\item{n}{integer. Number of observations.}

\item{lambda}{numeric. Mean.}

\item{d}{numeric. Dispersion parameter.}
}
\value{
a vector with the simulated sample.
}
\description{
Internal function to generate random numbers following over-dispersed
Poisson distribution based on simple cheat of using a standard negative
binomial, but choosing the scale parameter to give the desired mean vs
variance ratio at the given value of the mean.
Taken from: https://stat.ethz.ch/pipermail/r-help/2002-June/022425.html
}
\keyword{internal}
