% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatedlnm.R
\name{simulatedlnm}
\alias{simulatedlnm}
\alias{simulatedlnm.linearlme}
\alias{simulatedlnm.linearglm}
\alias{simulatedlnm.nonlinearglm}
\title{simulatedllm generic.}
\usage{
simulatedlnm(x)

\method{simulatedlnm}{linearlme}(x)

\method{simulatedlnm}{linearglm}(x)

\method{simulatedlnm}{nonlinearglm}(x)
}
\arguments{
\item{x}{an object internally generated by the \code{\link{collindlnm}} function.}
}
\description{
Simulation procedure internally called by the \code{\link{collindlnm}}
function for a given hypothetical effect.
}
\section{Methods (by class)}{
\itemize{
\item \code{simulatedlnm(linearlme)}: case of a hypothetical linear effect in a model of
class \code{"lme"}.

\item \code{simulatedlnm(linearglm)}: case of a hypothetical linear effect in a model of
class \code{"glm"}.

\item \code{simulatedlnm(nonlinearglm)}: case of a hypothetical non-linear effect in a model
of class \code{"glm"}.

}}
